Experiment moving ipykernel message handling to notebook
