+++ src/kernels/execution/cellExecutionMessageHandler.ts
@@ -43,9 +43,8 @@ import { swallowExceptions } from '../../platform/common/utils/decorators';
 import { noop } from '../../platform/common/utils/misc';
 import { ITracebackFormatter } from '../../kernels/types';
 import { handleTensorBoardDisplayDataOutput } from './executionHelpers';
-import { WIDGET_MIMETYPE } from '../ipywidgets/constants';
 import isObject = require('lodash/isObject');
-import { Identifiers } from '../../platform/common/constants';
+import { Identifiers, WIDGET_MIMETYPE } from '../../platform/common/constants';
 import { Lazy } from '../../platform/common/utils/lazy';
 
 // Helper interface for the set_next_input execute reply payload
