--- src/kernels/ipywidgets/constants.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export const WIDGET_MIMETYPE = 'application/vnd.jupyter.widget-view+json';
