+++ src/notebooks/controllers/ipywidgets/ipyWidgetScriptSourceProvider.ts
@@ -3,12 +3,12 @@
 
 'use strict';
 
-import '../../platform/common/extensions';
-import { traceError, traceInfo } from '../../platform/logging';
-import { WidgetCDNs, IConfigurationService, IHttpClient } from '../../platform/common/types';
-import { sendTelemetryEvent, Telemetry } from '../../telemetry';
-import { getTelemetrySafeHashedString } from '../../platform/telemetry/helpers';
-import { IKernel } from '../types';
+import '../../../platform/common/extensions';
+import { traceError, traceInfo } from '../../../platform/logging';
+import { WidgetCDNs, IConfigurationService, IHttpClient } from '../../../platform/common/types';
+import { sendTelemetryEvent, Telemetry } from '../../../telemetry';
+import { getTelemetrySafeHashedString } from '../../../platform/telemetry/helpers';
+import { IKernel } from '../../../kernels/types';
 import {
     ILocalResourceUriConverter,
     IWidgetScriptSourceProvider,
