+++ src/notebooks/controllers/ipywidgets/nbExtensionsPathProvider.node.ts
@@ -3,11 +3,11 @@
 
 import { injectable } from 'inversify';
 import { Uri } from 'vscode';
-import { IKernel } from '../types';
+import { IKernel } from '../../../kernels/types';
 import { INbExtensionsPathProvider } from './types';
 
 /**
- * Returns the path to the nbExtensions folder for a given kernel (web)
+ * Returns the path to the nbExtensions folder for a given kernel (node)
  */
 @injectable()
 export class NbExtensionsPathProvider implements INbExtensionsPathProvider {
@@ -17,8 +17,16 @@ export class NbExtensionsPathProvider implements INbExtensionsPathProvider {
             case 'startUsingRemoteKernelSpec': {
                 return Uri.parse(kernel.kernelConnectionMetadata.baseUrl);
             }
+            case 'startUsingPythonInterpreter': {
+                return Uri.joinPath(
+                    Uri.file(kernel.kernelConnectionMetadata.interpreter.sysPrefix),
+                    'share',
+                    'jupyter'
+                );
+            }
             default: {
-                // Not possible a possible code path in web.
+                // We haven't come across scenarios with non-python kernels that use widgets
+                // & have custom widget sources. If we do, we can implement that as we come across them.
                 return;
             }
         }
