+++ src/notebooks/controllers/ipywidgets/scriptUriConverter.ts
