+++ src/notebooks/controllers/ipywidgets/serviceRegistry.node.ts
@@ -1,9 +1,9 @@
-import { IServiceManager } from '../../platform/ioc/types';
-import { ScriptSourceProviderFactory } from './scriptSourceProviderFactory.web';
+import { IServiceManager } from '../../../platform/ioc/types';
+import { ScriptSourceProviderFactory } from './scriptSourceProviderFactory.node';
 import { IIPyWidgetScriptManagerFactory, INbExtensionsPathProvider, IWidgetScriptSourceProviderFactory } from './types';
 import { IPyWidgetMessageDispatcherFactory } from './ipyWidgetMessageDispatcherFactory';
-import { NbExtensionsPathProvider } from './nbExtensionsPathProvider.web';
-import { IPyWidgetScriptManagerFactory } from './ipyWidgetScriptManagerFactory.web';
+import { NbExtensionsPathProvider } from './nbExtensionsPathProvider.node';
+import { IPyWidgetScriptManagerFactory } from './ipyWidgetScriptManagerFactory.node';
 
 export function registerTypes(serviceManager: IServiceManager, _isDevMode: boolean) {
     serviceManager.addSingleton<IPyWidgetMessageDispatcherFactory>(
