+++ src/test/datascience/widgets/thirdpartyWidgets.vscode.common.test.ts
@@ -33,7 +33,7 @@ import {
     executeCellAndWaitForOutput,
     initializeNotebookForWidgetTest
 } from './standardWidgets.vscode.common.test';
-import { GlobalStateKeyToTrackIfUserConfiguredCDNAtLeastOnce } from '../../../kernels/ipywidgets/cdnWidgetScriptSourceProvider';
+import { GlobalStateKeyToTrackIfUserConfiguredCDNAtLeastOnce } from '../../../notebooks/controllers/ipywidgets/cdnWidgetScriptSourceProvider';
 
 [true, false].forEach((useCDN) => {
     /* eslint-disable @typescript-eslint/no-explicit-any, no-invalid-this */
