+++ src/webviews/webview-side/ipywidgets/common/incompatibleWidgetHandler.ts
@@ -4,7 +4,7 @@
 'use strict';
 
 import * as semver from 'semver';
-import { WidgetScriptSource } from '../../../../kernels/ipywidgets/types';
+import { WidgetScriptSource } from '../../../../notebooks/controllers/ipywidgets/types';
 const supportedVersionOfQgrid = '1.1.1';
 const qgridModuleName = 'qgrid';
