+++ src/webviews/webview-side/ipywidgets/common/manager.ts
@@ -16,10 +16,10 @@ import { logMessage, setLogger } from '../../react-common/logger';
 import { IMessageHandler, PostOffice } from '../../react-common/postOffice';
 import { create as createKernel } from './kernel';
 import { IIPyWidgetManager, IJupyterLabWidgetManager, IJupyterLabWidgetManagerCtor, ScriptLoader } from './types';
-import { WIDGET_MIMETYPE } from '../../../../kernels/ipywidgets/constants';
 import { KernelSocketOptions } from '../../../../kernels/types';
 import { Deferred, createDeferred } from '../../../../platform/common/utils/async';
 import { IInteractiveWindowMapping, IPyWidgetMessages, InteractiveWindowMessages } from '../../../../messageTypes';
+import { WIDGET_MIMETYPE } from '../../../../platform/common/constants';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
